Split K8s deploy tests into one class per file for CI parallelization
Each test class becomes a separate CI job via SplitTestsOnCI=true.
