+++ tests/Aspire.Cli.EndToEnd.Tests/KubernetesDeployWithMySqlTests.cs
@@ -1,0 +1,150 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using Aspire.Cli.EndToEnd.Tests.Helpers;
+using Aspire.Cli.Tests.Utils;
+using Hex1b.Automation;
+using Xunit;
+
+namespace Aspire.Cli.EndToEnd.Tests;
+
+/// <summary>
+/// E2E test for <c>aspire deploy</c> to Kubernetes: DeployWithMySql.
+/// </summary>
+public sealed class KubernetesDeployWithMySqlTests(ITestOutputHelper output)
+{
+    private const string ProjectName = "K8sDeployTest";
+
+    [Fact]
+    [CaptureWorkspaceOnFailure]
+    public async Task DeployWithMySql()
+    {
+        using var workspace = TemporaryWorkspace.Create(output);
+
+        var prNumber = CliE2ETestHelpers.GetRequiredPrNumber();
+        var commitSha = CliE2ETestHelpers.GetRequiredCommitSha();
+        var isCI = CliE2ETestHelpers.IsRunningInCI;
+        var clusterName = KubernetesDeployTestHelpers.GenerateUniqueClusterName();
+        var k8sNamespace = $"test-{clusterName[..16]}";
+
+        output.WriteLine($"Cluster name: {clusterName}");
+        output.WriteLine($"Namespace: {k8sNamespace}");
+
+        using var terminal = CliE2ETestHelpers.CreateTestTerminal();
+        var pendingRun = terminal.RunAsync(TestContext.Current.CancellationToken);
+
+        var counter = new SequenceCounter();
+        var auto = new Hex1bTerminalAutomator(terminal, defaultTimeout: TimeSpan.FromSeconds(500));
+
+        await auto.PrepareEnvironmentAsync(workspace, counter);
+
+        if (isCI)
+        {
+            await auto.InstallAspireCliFromPullRequestAsync(prNumber, counter);
+            await auto.SourceAspireCliEnvironmentAsync(counter);
+            await auto.VerifyAspireCliVersionAsync(commitSha, counter);
+        }
+
+        try
+        {
+            await auto.InstallKindAndHelmAsync(counter);
+            await auto.CreateKindClusterWithRegistryAsync(counter, clusterName);
+
+            var appHostCode = $$"""
+                using Aspire.Hosting;
+                using Aspire.Hosting.Kubernetes;
+
+                var builder = DistributedApplication.CreateBuilder(args);
+
+                var registryEndpoint = builder.AddParameter("registryendpoint");
+                var registry = builder.AddContainerRegistry("registry", registryEndpoint);
+
+                var mysql = builder.AddMySql("mysql").AddDatabase("testdb");
+
+                var api = builder.AddProject<Projects.{{ProjectName}}_ApiService>("server")
+                    .WithReference(mysql)
+                    .WaitFor(mysql)
+                    .WithExternalHttpEndpoints();
+
+                builder.AddKubernetesEnvironment("env")
+                    .WithHelm(helm =>
+                    {
+                        helm.WithNamespace(builder.AddParameter("namespace"));
+                        helm.WithChartVersion(builder.AddParameter("chartversion"));
+                    });
+
+                builder.Build().Run();
+                """;
+
+            var apiProgramCode = """
+                using MySqlConnector;
+
+                var builder = WebApplication.CreateBuilder(args);
+                builder.AddServiceDefaults();
+                builder.AddMySqlDataSource("testdb");
+
+                var app = builder.Build();
+                app.MapDefaultEndpoints();
+
+                app.MapGet("/test-deployment", async (MySqlDataSource dataSource) =>
+                {
+                    await using var conn = await dataSource.OpenConnectionAsync();
+                    await using var cmd = conn.CreateCommand();
+                    cmd.CommandText = "SELECT 1 AS result";
+                    var result = await cmd.ExecuteScalarAsync();
+
+                    if (result is int val && val == 1)
+                    {
+                        return Results.Ok("PASSED: MySQL SELECT 1 works");
+                    }
+                    return Results.Problem($"FAILED: expected 1, got '{result}'");
+                });
+
+                app.Run();
+                """;
+
+            KubernetesDeployTestHelpers.ScaffoldK8sDeployProject(
+                workspace.WorkspaceRoot.FullName,
+                ProjectName,
+                appHostHostingPackages: ["Aspire.Hosting.Kubernetes", "Aspire.Hosting.MySql"],
+                apiClientPackages: ["Aspire.MySqlConnector"],
+                appHostCode: appHostCode,
+                apiProgramCode: apiProgramCode);
+
+            await auto.TypeAsync($"cd {ProjectName}");
+            await auto.EnterAsync();
+            await auto.WaitForSuccessPromptAsync(counter);
+
+            await auto.TypeAsync("unset ASPIRE_PLAYGROUND");
+            await auto.EnterAsync();
+            await auto.WaitForSuccessPromptAsync(counter);
+
+            await auto.AspireDeployInteractiveAsync(
+                counter,
+                parameterResponses:
+                [
+                    ("chartversion", "0.1.0"),
+                    ("namespace", k8sNamespace),
+                    ("registryendpoint", "localhost:5001"),
+                ]);
+
+            await auto.VerifyDeploymentAsync(
+                counter,
+                @namespace: k8sNamespace,
+                serviceName: "server",
+                localPort: 18085,
+                testPath: "/test-deployment");
+
+            await auto.CleanupKubernetesDeploymentAsync(counter, clusterName);
+
+            await auto.TypeAsync("exit");
+            await auto.EnterAsync();
+        }
+        finally
+        {
+            await KubernetesDeployTestHelpers.CleanupKindClusterOutOfBandAsync(clusterName, output);
+        }
+
+        await pendingRun;
+    }
+}
